Final de la Tercera Iteracion 99.9999%

+++ src/main/java/com/TP20192/SRVTFL/models/dao/IPreguntasDao.java
@@ -25,5 +25,5 @@ public interface IPreguntasDao extends PagingAndSortingRepository<Pregunta, Long
             + " where p.pregPrimeraVez= :preP and f.fobId = :fobId",
             countQuery = "select count(p) from Pregunta p join p.fobId f"
                     + " where p.pregPrimeraVez = :preP and f.fobId = :fobId")
-    public Page<Pregunta> EncontrarPreguntasCita(@Param("preP") Boolean preP,@Param("fobId") int fobId ,Pageable pageable);
+    public Page<Pregunta> EncontrarPreguntasCita(@Param("preP") Boolean preP,@Param("fobId") Long fobId ,Pageable pageable);
 }
