+++ src/main/java/com/TP20192/SRVTFL/models/service/ICitaService.java
@@ -74,7 +74,7 @@ public interface ICitaService {
    public Cita encontrarCitaconPacinenteconEstado(Long id);
    
    //
-   public Page<Pregunta> EncontrarPreguntasCita(Boolean tratId,int fobId, Pageable pageable);
+   public Page<Pregunta> EncontrarPreguntasCita(Boolean tratId,Long fobId, Pageable pageable);
    
    public Pregunta encontrarPregunta(Long Id);
